corretto errore critico LCS, edit suppongo funzioni ora (?)

+++ clustering/lcs/EditDistance.h
@@ -8,9 +8,9 @@
 #ifndef CLUSTERING_LCS_EDITDISTANCE_H_
 #define CLUSTERING_LCS_EDITDISTANCE_H_
 
-#include <stddef.h>
+
 #include <string>
 
-size_t edit_distance(const std::string& A, const std::string& B);
+int edit_distance(const std::string& A, const std::string& B);
 
 #endif /* CLUSTERING_LCS_EDITDISTANCE_H_ */
